requirements.txt
torch==1.12.1
torchvision==0.13.1
pyyaml==6.0.1
pytorch3d==0.7.5
trimesh==4.1.3
scikit-learn==1.3.0
einops==0.7.0
potpourri3d==1.0.0
